+++ src/ui/promgr/promgr.go
@@ -15,22 +15,229 @@
 package promgr
 
 import (
+	"fmt"
+	"strconv"
+
 	"github.com/vmware/harbor/src/common/models"
+	"github.com/vmware/harbor/src/common/utils/log"
+	"github.com/vmware/harbor/src/ui/promgr/metamgr"
+	"github.com/vmware/harbor/src/ui/promgr/pmsdriver"
 )
 
-// ProMgr is the project mamager which abstracts the operations related
+// ProjectManager is the project mamager which abstracts the operations related
 // to projects
-type ProMgr interface {
+type ProjectManager interface {
 	Get(projectIDOrName interface{}) (*models.Project, error)
-	IsPublic(projectIDOrName interface{}) (bool, error)
-	Exist(projectIDOrName interface{}) (bool, error)
-	// get all public project
-	GetPublic() ([]*models.Project, error)
 	Create(*models.Project) (int64, error)
 	Delete(projectIDOrName interface{}) error
 	Update(projectIDOrName interface{}, project *models.Project) error
-	// GetAll returns a project list according to the query parameters
-	GetAll(query *models.ProjectQueryParam, base ...*models.BaseProjectCollection) ([]*models.Project, error)
-	// GetTotal returns the total count according to the query parameters
-	GetTotal(query *models.ProjectQueryParam, base ...*models.BaseProjectCollection) (int64, error)
+	List(query *models.ProjectQueryParam) (*models.ProjectQueryResult, error)
+	IsPublic(projectIDOrName interface{}) (bool, error)
+	Exists(projectIDOrName interface{}) (bool, error)
+	// get all public project
+	GetPublic() ([]*models.Project, error)
+	// if the project manager uses a metadata manager, return it, otherwise return nil
+	GetMetadataManager() metamgr.ProjectMetadataManager
+}
+
+type defaultProjectManager struct {
+	pmsDriver      pmsdriver.PMSDriver
+	metaMgrEnabled bool // if metaMgrEnabled is enabled, metaMgr will be used to CURD metadata
+	metaMgr        metamgr.ProjectMetadataManager
+}
+
+// NewDefaultProjectManager returns an instance of defaultProjectManager,
+// if metaMgrEnabled is true, a project metadata manager will be created
+// and used to CURD metadata
+func NewDefaultProjectManager(driver pmsdriver.PMSDriver, metaMgrEnabled bool) ProjectManager {
+	mgr := &defaultProjectManager{
+		pmsDriver:      driver,
+		metaMgrEnabled: metaMgrEnabled,
+	}
+	if metaMgrEnabled {
+		mgr.metaMgr = metamgr.NewDefaultProjectMetadataManager()
+	}
+	return mgr
+}
+
+func (d *defaultProjectManager) Get(projectIDOrName interface{}) (*models.Project, error) {
+	project, err := d.pmsDriver.Get(projectIDOrName)
+	if err != nil {
+		return nil, err
+	}
+
+	if project != nil && d.metaMgrEnabled {
+		meta, err := d.metaMgr.Get(project.ProjectID)
+		if err != nil {
+			return nil, err
+		}
+		if len(project.Metadata) == 0 {
+			project.Metadata = make(map[string]string)
+		}
+		for k, v := range meta {
+			project.Metadata[k] = v
+		}
+	}
+	return project, nil
+}
+func (d *defaultProjectManager) Create(project *models.Project) (int64, error) {
+	id, err := d.pmsDriver.Create(project)
+	if err != nil {
+		return 0, err
+	}
+	if len(project.Metadata) > 0 && d.metaMgrEnabled {
+		if err = d.metaMgr.Add(id, project.Metadata); err != nil {
+			log.Errorf("failed to add metadata for project %s: %v", project.Name, err)
+		}
+	}
+	return id, nil
+}
+
+func (d *defaultProjectManager) Delete(projectIDOrName interface{}) error {
+	project, err := d.Get(projectIDOrName)
+	if err != nil {
+		return err
+	}
+	if project == nil {
+		return nil
+	}
+	if project.Metadata != nil && d.metaMgrEnabled {
+		if err = d.metaMgr.Delete(project.ProjectID); err != nil {
+			return err
+		}
+	}
+	return d.pmsDriver.Delete(project.ProjectID)
+}
+
+func (d *defaultProjectManager) Update(projectIDOrName interface{}, project *models.Project) error {
+	if len(project.Metadata) > 0 && d.metaMgrEnabled {
+		pro, err := d.Get(projectIDOrName)
+		if err != nil {
+			return err
+		}
+		if pro == nil {
+			return fmt.Errorf("project %v not found", projectIDOrName)
+		}
+
+		// TODO transaction?
+		metaNeedUpdated := map[string]string{}
+		metaNeedCreated := map[string]string{}
+		if pro.Metadata == nil {
+			pro.Metadata = map[string]string{}
+		}
+		for key, value := range project.Metadata {
+			_, exist := pro.Metadata[key]
+			if exist {
+				metaNeedUpdated[key] = value
+			} else {
+				metaNeedCreated[key] = value
+			}
+		}
+		if err = d.metaMgr.Add(pro.ProjectID, metaNeedCreated); err != nil {
+			return err
+		}
+		if err = d.metaMgr.Update(pro.ProjectID, metaNeedUpdated); err != nil {
+			return err
+		}
+	}
+
+	return d.pmsDriver.Update(projectIDOrName, project)
+}
+
+func (d *defaultProjectManager) List(query *models.ProjectQueryParam) (*models.ProjectQueryResult, error) {
+	// query by public/private property with ProjectMetadataManager first
+	if d.metaMgrEnabled && query != nil && query.Public != nil {
+		projectIDs, err := d.filterByPublic(*query.Public)
+		if err != nil {
+			return nil, err
+		}
+
+		if len(projectIDs) == 0 {
+			return &models.ProjectQueryResult{}, nil
+		}
+
+		if query.ProjectIDs == nil {
+			query.ProjectIDs = projectIDs
+		} else {
+			query.ProjectIDs = findInBoth(query.ProjectIDs, projectIDs)
+		}
+	}
+
+	// query by other properties
+	result, err := d.pmsDriver.List(query)
+	if err != nil {
+		return nil, err
+	}
+
+	// populate metadata
+	if d.metaMgrEnabled {
+		for _, project := range result.Projects {
+			meta, err := d.metaMgr.Get(project.ProjectID)
+			if err != nil {
+				return nil, err
+			}
+			project.Metadata = meta
+		}
+	}
+	return result, nil
+}
+
+func (d *defaultProjectManager) filterByPublic(public bool) ([]int64, error) {
+	metas, err := d.metaMgr.List(models.ProMetaPublic, strconv.FormatBool(public))
+	if err != nil {
+		return nil, err
+	}
+
+	projectIDs := []int64{}
+	for _, meta := range metas {
+		projectIDs = append(projectIDs, meta.ProjectID)
+	}
+	return projectIDs, nil
+}
+
+func findInBoth(ids1 []int64, ids2 []int64) []int64 {
+	m := map[int64]struct{}{}
+	for _, id := range ids1 {
+		m[id] = struct{}{}
+	}
+
+	ids := []int64{}
+	for _, id := range ids2 {
+		if _, exist := m[id]; exist {
+			ids = append(ids, id)
+		}
+	}
+
+	return ids
+}
+
+func (d *defaultProjectManager) IsPublic(projectIDOrName interface{}) (bool, error) {
+	project, err := d.Get(projectIDOrName)
+	if err != nil {
+		return false, err
+	}
+	if project == nil {
+		return false, nil
+	}
+	return project.IsPublic(), nil
+}
+
+func (d *defaultProjectManager) Exists(projectIDOrName interface{}) (bool, error) {
+	project, err := d.Get(projectIDOrName)
+	return project != nil, err
+}
+
+func (d *defaultProjectManager) GetPublic() ([]*models.Project, error) {
+	value := true
+	result, err := d.List(&models.ProjectQueryParam{
+		Public: &value,
+	})
+	if err != nil {
+		return nil, err
+	}
+	return result.Projects, nil
+}
+
+func (d *defaultProjectManager) GetMetadataManager() metamgr.ProjectMetadataManager {
+	return d.metaMgr
 }